Refactor integration test code
