+++ src/integration-tests/vultr-noop_test.ts
@@ -1,48 +1,38 @@
-import { VultrProvider } from "../providers/vultr/mod.ts";
-import { VultrApiMock } from "../providers/vultr/mock.ts";
-import { applyToProvider } from "../common/test-utils.ts";
-import { NoopRegistry } from "../registries/noop.ts";
+import { mockedVultrTest } from "./vultr-mock.ts";
 
-Deno.test("[E2E: Vultr & noop] Update changed A record", async () => {
-
-  const registry = new NoopRegistry({
-    type: 'noop',
-  });
-
-  const apiMock = new VultrApiMock();
-  apiMock.addMockedZone('example.com', [{
+Deno.test("[E2E: Vultr & noop] Update changed A record",
+  async () => await mockedVultrTest({
+    registry: { type: 'noop', },
+    sourceRecords: [{
+      annotations: {},
+      resourceKey: 'e2e',
+      dns: {
+        fqdn: "www.example.com",
+        type: "A",
+        target: '2.2.2.2',
+      },
+    }],
+  }).withZone('example.com', [{
     expect: 'deletion',
     data: { name: 'www', type: 'A', data: '1.1.1.1' },
   }, {
     expect: 'creation',
     data: { name: 'www', type: 'A', data: '2.2.2.2' },
-  }]);
+  }]).go());
 
-  const provider = new VultrProvider({
-    type: 'vultr',
-  }, apiMock);
-
-  await applyToProvider(provider, registry, [{
-    annotations: {},
-    resourceKey: 'e2e',
-    dns: {
-      fqdn: "www.example.com",
-      type: "A",
-      target: '2.2.2.2',
-    },
-  }]);
-
-  apiMock.verifyCompletion();
-});
-
-Deno.test("[E2E: Vultr & noop] Clean up unrelated records", async () => {
-
-  const registry = new NoopRegistry({
-    type: 'noop',
-  });
-
-  const apiMock = new VultrApiMock();
-  apiMock.addMockedZone('example.com', [{
+Deno.test("[E2E: Vultr & noop] Clean up unrelated records",
+  async () => await mockedVultrTest({
+    registry: { type: 'noop', },
+    sourceRecords: [{
+      annotations: {},
+      resourceKey: 'e2e',
+      dns: {
+        fqdn: "www.example.com",
+        type: "A",
+        target: '1.1.1.1',
+      },
+    }],
+  }).withZone('example.com', [{
     // noop registry shouldn't try removing the zone's identity records
     expect: 'retained',
     data: { name: '', type: 'NS', data: 'ns1.website' },
@@ -58,21 +48,4 @@ Deno.test("[E2E: Vultr & noop] Clean up unrelated records", async () => {
   }, {
     expect: 'deletion',
     data: { name: '', type: 'MX', data: '5 mx.website' },
-  }]);
-
-  const provider = new VultrProvider({
-    type: 'vultr',
-  }, apiMock);
-
-  await applyToProvider(provider, registry, [{
-    annotations: {},
-    resourceKey: 'e2e',
-    dns: {
-      fqdn: "www.example.com",
-      type: "A",
-      target: '1.1.1.1',
-    },
-  }]);
-
-  apiMock.verifyCompletion();
-});
+  }]).go());
