+++ src/integration-tests/vultr-txt_test.ts
@@ -1,12 +1,14 @@
-import { VultrProvider } from "../providers/vultr/mod.ts";
-import { VultrApiMock } from "../providers/vultr/mock.ts";
-import { TxtRegistry } from "../registries/txt.ts";
-import { applyToProvider } from "../common/test-utils.ts";
-import { SourceRecord } from "../common/contract.ts";
-import type { TxtRegistryConfig, VultrProviderConfig } from "../common/config.ts";
+import { RegistryConfig } from "../common/config.ts";
+import { mockedVultrTest } from "./vultr-mock.ts";
+
+const defaultRegistry: RegistryConfig = {
+  type: 'txt',
+  txt_owner_id: 'dnssynctest',
+};
 
 Deno.test("[E2E: Vultr & TXT] Keep unchanged A record",
   async () => await mockedVultrTest({
+    registry: defaultRegistry,
     sourceRecords: [{
       annotations: {},
       resourceKey: 'e2e',
@@ -27,6 +29,7 @@ Deno.test("[E2E: Vultr & TXT] Keep unchanged A record",
 
 Deno.test("[E2E: Vultr & TXT] Update unscoped registry record",
   async () => await mockedVultrTest({
+    registry: defaultRegistry,
     sourceRecords: [{
       annotations: {},
       resourceKey: 'e2e',
@@ -51,6 +54,7 @@ Deno.test("[E2E: Vultr & TXT] Update unscoped registry record",
 
 Deno.test("[E2E: Vultr & TXT] Update outdated registry record",
   async () => await mockedVultrTest({
+    registry: defaultRegistry,
     sourceRecords: [{
       annotations: {},
       resourceKey: 'e2e',
@@ -75,7 +79,7 @@ Deno.test("[E2E: Vultr & TXT] Update outdated registry record",
 
 Deno.test("[E2E: Vultr & TXT] Update changed A record",
   async () => await mockedVultrTest({
-    registry: {
+    registry: { ...defaultRegistry,
       txt_prefix: 'registry.',
     },
     sourceRecords: [{
@@ -101,6 +105,7 @@ Deno.test("[E2E: Vultr & TXT] Update changed A record",
 
 Deno.test("[E2E: Vultr & TXT] Remove our abandoned A record",
   async () => await mockedVultrTest({
+    registry: defaultRegistry,
     sourceRecords: [],
   }).withZone('example.com', [{
     expect: 'deletion',
@@ -113,6 +118,7 @@ Deno.test("[E2E: Vultr & TXT] Remove our abandoned A record",
 
 Deno.test("[E2E: Vultr & TXT] Leave unmanaged records alone",
   async () => await mockedVultrTest({
+    registry: defaultRegistry,
     sourceRecords: [],
   }).withZone('example.com', [{
     expect: 'retained',
@@ -128,6 +134,7 @@ Deno.test("[E2E: Vultr & TXT] Leave unmanaged records alone",
 
 Deno.test("[E2E: Vultr & TXT] Be v careful around registry records",
   async () => await mockedVultrTest({
+    registry: defaultRegistry,
     sourceRecords: [],
   }).withZone('example.com', [{
     expect: 'retained',
@@ -144,6 +151,7 @@ Deno.test("[E2E: Vultr & TXT] Be v careful around registry records",
 
 Deno.test("[E2E: Vultr & TXT] Add new A record",
   async () => await mockedVultrTest({
+    registry: defaultRegistry,
     sourceRecords: [{
       annotations: {e2e: 'test'},
       resourceKey: 'e2e',
@@ -164,7 +172,7 @@ Deno.test("[E2E: Vultr & TXT] Add new A record",
 
 Deno.test("[E2E: Vultr & TXT] Adopt record from other owner",
   async () => await mockedVultrTest({
-    registry: {
+    registry: { ...defaultRegistry,
       txt_prefix: 'registry.',
       auto_adopt_from_owner_ids: ['legacy'],
     },
@@ -195,7 +203,7 @@ Deno.test("[E2E: Vultr & TXT] Adopt record from other owner",
 
 Deno.test("[E2E: Vultr & TXT] Adopt without changes",
   async () => await mockedVultrTest({
-    registry: {
+    registry: { ...defaultRegistry,
       txt_prefix: 'registry.',
       auto_adopt_from_owner_ids: ['legacy'],
     },
@@ -220,40 +228,3 @@ Deno.test("[E2E: Vultr & TXT] Adopt without changes",
     data: { name: 'registry.www', type: 'TXT', data:
       '"heritage=external-dns,external-dns/owner=dnssynctest,external-dns/resource=e2e,record-type/A=managed"' },
   }]).go());
-
-//////
-
-function mockedVultrTest(opts: {
-  registry?: Partial<TxtRegistryConfig>,
-  provider?: Partial<VultrProviderConfig>,
-  vultr?: VultrApiMock,
-  // vultrRecords?: DnsRecordData[],
-  sourceRecords: Array<SourceRecord>,
-}) {
-
-  const registry = new TxtRegistry({
-    type: 'txt',
-    txt_owner_id: 'dnssynctest',
-    ...opts.registry,
-  });
-  const apiMock = opts.vultr ?? new VultrApiMock();
-  const provider = new VultrProvider({
-    type: 'vultr',
-    ...opts.provider,
-  }, apiMock);
-
-  return {
-    withZone(...a: Parameters<typeof apiMock.addMockedZone>) {
-      apiMock.addMockedZone(...a);
-      return this;
-    },
-    withDeadZone(...a: Parameters<typeof apiMock.addMockedDeadZone>) {
-      apiMock.addMockedDeadZone(...a);
-      return this;
-    },
-    async go() {
-      await applyToProvider(provider, registry, opts.sourceRecords);
-      apiMock.verifyCompletion();
-    },
-  };
-}
