+++ src/providers/vultr/functional-test.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env -S deno test --allow-env --allow-net=api.vultr.com
 
 import { assertEquals, assertObjectMatch } from "https://deno.land/std@0.105.0/testing/asserts.ts";
-import { applyToProvider } from "../../common/test-utils.ts";
+import { applyToProvider } from "../../integration-tests/apply.ts";
 import { NoopRegistry } from "../../registries/noop.ts";
 import { VultrProvider } from "./mod.ts";
